remove v1alpha5 metrics and counter
