+++ pkg/controllers/controllers.go
@@ -25,7 +25,6 @@ import (
 	metricsnode "github.com/aws/karpenter-core/pkg/controllers/metrics/node"
 	metricsnodepool "github.com/aws/karpenter-core/pkg/controllers/metrics/nodepool"
 	metricspod "github.com/aws/karpenter-core/pkg/controllers/metrics/pod"
-	metricsprovisioner "github.com/aws/karpenter-core/pkg/controllers/metrics/provisioner"
 	"github.com/aws/karpenter-core/pkg/controllers/node/termination"
 	"github.com/aws/karpenter-core/pkg/controllers/node/termination/terminator"
 	nodeclaimconsistency "github.com/aws/karpenter-core/pkg/controllers/nodeclaim/consistency"
@@ -69,10 +68,8 @@ func NewControllers(
 		informer.NewNodeClaimController(kubeClient, cluster),
 		termination.NewController(kubeClient, cloudProvider, terminator.NewTerminator(clock, kubeClient, evictionQueue), recorder),
 		metricspod.NewController(kubeClient),
-		metricsprovisioner.NewController(kubeClient),
 		metricsnodepool.NewController(kubeClient),
 		metricsnode.NewController(cluster),
-		nodepoolcounter.NewProvisionerController(kubeClient, cluster),
 		nodepoolcounter.NewNodePoolController(kubeClient, cluster),
 		nodeclaimconsistency.NewMachineController(clock, kubeClient, recorder, cloudProvider),
 		nodeclaimconsistency.NewNodeClaimController(clock, kubeClient, recorder, cloudProvider),
