+++ pkg/controllers/nodepool/counter/controller.go
@@ -131,46 +131,3 @@ func (c *NodePoolController) Builder(_ context.Context, m manager.Manager) corec
 		).
 		WithOptions(controller.Options{MaxConcurrentReconciles: 10}))
 }
-
-type ProvisionerController struct {
-	*Controller
-}
-
-func NewProvisionerController(kubeClient client.Client, cluster *state.Cluster) corecontroller.Controller {
-	return corecontroller.Typed[*v1alpha5.Provisioner](kubeClient, &ProvisionerController{
-		Controller: NewController(kubeClient, cluster),
-	})
-}
-
-func (c *ProvisionerController) Reconcile(ctx context.Context, provisioner *v1alpha5.Provisioner) (reconcile.Result, error) {
-	return c.Controller.Reconcile(ctx, nodepoolutil.New(provisioner))
-}
-
-func (c *ProvisionerController) Name() string {
-	return "counter"
-}
-
-func (c *ProvisionerController) Builder(_ context.Context, m manager.Manager) corecontroller.Builder {
-	return corecontroller.Adapt(controllerruntime.
-		NewControllerManagedBy(m).
-		For(&v1alpha5.Provisioner{}).
-		Watches(
-			&v1alpha5.Machine{},
-			handler.EnqueueRequestsFromMapFunc(func(_ context.Context, o client.Object) []reconcile.Request {
-				if name, ok := o.GetLabels()[v1alpha5.ProvisionerNameLabelKey]; ok {
-					return []reconcile.Request{{NamespacedName: types.NamespacedName{Name: name}}}
-				}
-				return nil
-			}),
-		).
-		Watches(
-			&v1.Node{},
-			handler.EnqueueRequestsFromMapFunc(func(_ context.Context, o client.Object) []reconcile.Request {
-				if name, ok := o.GetLabels()[v1alpha5.ProvisionerNameLabelKey]; ok {
-					return []reconcile.Request{{NamespacedName: types.NamespacedName{Name: name}}}
-				}
-				return nil
-			}),
-		).
-		WithOptions(controller.Options{MaxConcurrentReconciles: 10}))
-}
